Split screen to small components

+++ lib/movies/data/models/movie_model.dart
@@ -15,7 +15,6 @@ class MovieModel extends Movie {
       backdropPath: json["backdrop_path"],
       genreIds: List<int>.from(json["genre_ids"].map((e) => e)),
       overview: json["overview"],
-      //TODO : CHECK THIS
-      voteAverage: json["vote_average"],
+      voteAverage: json["vote_average"].toDouble(),
       relaesDate: json["release_date"]);
 }
